feat: accept decimal quantities on line items
Line item quantity widens from integer to decimal(15,3) so a product sold
by weight or volume can be quoted, ordered, invoiced, delivered and
purchase-ordered at its real quantity. Both editable quantity inputs were
bare number fields defaulting to step=1, so 3.5 Kg could not be entered.

The Order to Invoice and Order to Delivery forms picked their quantity
from a <select> built by an integer loop over the remainder, so an order
line of 2.5 could only be invoiced as 2 - leaving 0.5 outstanding
forever. That dropdown was also the only guard against over-invoicing
and it lived entirely in the browser, so the cap is now recomputed and
enforced server-side.

Also fixes three float comparisons that were latent while quantities
were exact integers and become user-visible with decimals: the
orderComplete/invoiceComplete/deliveryComplete `> 0` checks, CheckAmount's
exact `==`, and the (int) cast that silently under-charged fractional
lines.

Breaking: the API `quantity` field is now a JSON number rather than an
integer. The request side is a widening - every previously valid payload
still passes. See docs/upgrading.md for the six ALTERs.

+++ src/Http/Resources/Api/V2/InvoiceLineResource.php
@@ -13,7 +13,7 @@ class InvoiceLineResource extends JsonResource
             'id' => $this->external_id,
             'product_id' => $this->whenLoaded('product', fn () => $this->product?->external_id),
             'product' => $this->whenLoaded('product', fn () => $this->product ? new ProductBriefResource($this->product) : null),
-            'quantity' => $this->quantity !== null ? (int) $this->quantity : null,
+            'quantity' => $this->quantity !== null ? (float) $this->quantity : null,
             'unit_price' => $this->price !== null ? $this->price / 100 : null,
             'amount' => $this->amount !== null ? $this->amount / 100 : null,
             'tax_rate' => $this->tax_rate !== null ? (float) $this->tax_rate : null,
